+++ src/utils.ts
@@ -1,4 +1,5 @@
 import BigNumber from 'bignumber.js';
+import { Reward } from './entities';
 
 type BigNumberish = number | string | BigNumber;
 
@@ -33,3 +34,77 @@ export const weiToGweiNumber = (weiAmount: BigNumberish) => {
   const gweiAmount = divide(weiAmount, ethUnits.gwei).toNumber();
   return gweiAmount;
 };
+
+export const samplingCurve = (
+  sumWeight: number,
+  sampleMin: number,
+  sampleMax: number
+) => {
+  if (sumWeight <= sampleMin) {
+    return 0;
+  }
+  if (sumWeight >= sampleMax) {
+    return 1;
+  }
+  return (
+    (1 -
+      Math.cos(
+        ((sumWeight - sampleMin) * 2 * Math.PI) / (sampleMax - sampleMin)
+      )) /
+    2
+  );
+};
+
+export const linearRegression = (y: number[], x: number[]) => {
+  const n = y.length;
+  let sumX = 0;
+  let sumY = 0;
+  let sumXY = 0;
+  let sumXX = 0;
+
+  for (let i = 0; i < y.length; i++) {
+    const cY = Number(y[i]);
+    const cX = Number(x[i]);
+    sumX += cX;
+    sumY += cY;
+    sumXY += cX * cY;
+    sumXX += cX * cX;
+  }
+  const slope = (n * sumXY - sumX * sumY) / (n * sumXX - sumX * sumX);
+
+  return slope;
+};
+
+export const suggestBaseFee = (
+  baseFee: number[],
+  order: number[],
+  timeFactor: number,
+  sampleMin: number,
+  sampleMax: number
+) => {
+  if (timeFactor < 1e-6) {
+    return baseFee[baseFee.length - 1];
+  }
+  const pendingWeight =
+    (1 - Math.exp(-1 / timeFactor)) /
+    (1 - Math.exp(-baseFee.length / timeFactor));
+  let sumWeight = 0;
+  let result = 0;
+  let samplingCurveLast = 0;
+  for (let or of order) {
+    sumWeight +=
+      pendingWeight * Math.exp((or - baseFee.length + 1) / timeFactor);
+    const samplingCurveValue = samplingCurve(sumWeight, sampleMin, sampleMax);
+    result += (samplingCurveValue - samplingCurveLast) * baseFee[or];
+    if (samplingCurveValue >= 1) {
+      return result;
+    }
+    samplingCurveLast = samplingCurveValue;
+  }
+  return result;
+};
+
+export const rewardsFilterOutliers = (blocksRewards: Reward[], index: number) =>
+  blocksRewards
+    .map((reward) => weiToGweiNumber(reward[index]))
+    .filter((gweiReward) => gweiReward <= 10);
